Server -> SocketMap + Server
